+++ frontend/src/components/Footer/Footer.tsx
@@ -1,8 +1,8 @@
-import './Footer.module.css';
+import styles from './Footer.module.css';
 
 export default function Footer() {
     return (
-        <footer>
+        <footer className={styles['app-footer']}>
             <p>© 2025 Asset Management</p>
         </footer>
     );